+++ src/components/Avatar/Avatar.tsx
@@ -12,15 +12,11 @@ export interface IAvatarProps {
     type?: 'circle' | 'square';
 }
 
-export const Avatar = ({
-    src = Empty,
-    size = 38,
-    type = 'square',
-}: IAvatarProps) => (
+export const Avatar = ({ src, size = 38, type = 'square' }: IAvatarProps) => (
     <img
         width={size}
         height={size}
         className={`unique-avatar ${type}`}
-        src={src}
+        src={src?.length ? src : Empty}
     />
 );